feat: manage one event for gitlab
