+++ engine/hooks/tasks_test.go
@@ -86,102 +86,6 @@ func Test_doWebHookExecutionTagGithub(t *testing.T) {
 	assert.Equal(t, "0d1a26e67d8f5eaf1f6ba5c57fc3c7d91ac0fd1c", hs[0].Payload["git.hash"])
 }
 
-func Test_doWebHookExecutionGitlab(t *testing.T) {
-	log.SetLogger(t)
-	s, cancel := setupTestHookService(t)
-	defer cancel()
-	task := &sdk.TaskExecution{
-		UUID: sdk.RandomString(10),
-		Type: TypeRepoManagerWebHook,
-		WebHook: &sdk.WebHookExecution{
-			RequestBody: []byte(gitlabPushEvent),
-			RequestHeader: map[string][]string{
-				GitlabHeader: {"Push Hook"},
-			},
-			RequestURL: "",
-		},
-	}
-	hs, err := s.doWebHookExecution(task)
-	test.NoError(t, err)
-
-	assert.Equal(t, 1, len(hs))
-	assert.Equal(t, "master", hs[0].Payload["git.branch"])
-	assert.Equal(t, "jsmith", hs[0].Payload["git.author"])
-	assert.Equal(t, "Update Catalan translation to e38cb41.", hs[0].Payload["git.message"])
-	assert.Equal(t, "da1560886d4f094c3e6c9ef40349f7d38b5d27d7", hs[0].Payload["git.hash"])
-}
-
-var gitlabPushEvent = `
-	{
-  "object_kind": "push",
-  "before": "95790bf891e76fee5e1747ab589903a6a1f80f22",
-  "after": "da1560886d4f094c3e6c9ef40349f7d38b5d27d7",
-  "ref": "refs/heads/master",
-  "checkout_sha": "da1560886d4f094c3e6c9ef40349f7d38b5d27d7",
-  "user_id": 4,
-  "user_name": "John Smith",
-  "user_username": "jsmith",
-  "user_email": "john@example.com",
-  "user_avatar": "https://s.gravatar.com/avatar/d4c74594d841139328695756648b6bd6?s=8://s.gravatar.com/avatar/d4c74594d841139328695756648b6bd6?s=80",
-  "project_id": 15,
-  "project":{
-    "id": 15,
-    "name":"Diaspora",
-    "description":"",
-    "web_url":"http://example.com/mike/diaspora",
-    "avatar_url":null,
-    "git_ssh_url":"git@example.com:mike/diaspora.git",
-    "git_http_url":"http://example.com/mike/diaspora.git",
-    "namespace":"Mike",
-    "visibility_level":0,
-    "path_with_namespace":"mike/diaspora",
-    "default_branch":"master",
-    "homepage":"http://example.com/mike/diaspora",
-    "url":"git@example.com:mike/diaspora.git",
-    "ssh_url":"git@example.com:mike/diaspora.git",
-    "http_url":"http://example.com/mike/diaspora.git"
-  },
-  "repository":{
-    "name": "Diaspora",
-    "url": "git@example.com:mike/diaspora.git",
-    "description": "",
-    "homepage": "http://example.com/mike/diaspora",
-    "git_http_url":"http://example.com/mike/diaspora.git",
-    "git_ssh_url":"git@example.com:mike/diaspora.git",
-    "visibility_level":0
-  },
-  "commits": [
-    {
-      "id": "b6568db1bc1dcd7f8b4d5a946b0b91f9dacd7327",
-      "message": "Update Catalan translation to e38cb41.",
-      "timestamp": "2011-12-12T14:27:31+02:00",
-      "url": "http://example.com/mike/diaspora/commit/b6568db1bc1dcd7f8b4d5a946b0b91f9dacd7327",
-      "author": {
-        "name": "Jordi Mallach",
-        "email": "jordi@softcatala.org"
-      },
-      "added": ["CHANGELOG"],
-      "modified": ["app/controller/application.rb"],
-      "removed": []
-    },
-    {
-      "id": "da1560886d4f094c3e6c9ef40349f7d38b5d27d7",
-      "message": "fixed readme",
-      "timestamp": "2012-01-03T23:36:29+02:00",
-      "url": "http://example.com/mike/diaspora/commit/da1560886d4f094c3e6c9ef40349f7d38b5d27d7",
-      "author": {
-        "name": "GitLab dev user",
-        "email": "gitlabdev@dv6700.(none)"
-      },
-      "added": ["CHANGELOG"],
-      "modified": ["app/controller/application.rb"],
-      "removed": []
-    }
-  ],
-  "total_commits_count": 4
-}
-`
-
 var githubPushEvent = `
 	{
   "ref": "refs/heads/my-branch",
